Update training, data processing and cli
This change adds a procedure to select optimal alpha value and model
during training as well as reporting the perplexity of the resulting
model.
A few interventions to clean up and standardise data processing
code as well as minor additions to command line interface for better
feedback to the user.